vr_ml_transfer-develop/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
Brotli==1.0.9
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.104.1
fastjsonschema==2.18.1
faust-cchardet==2.1.19
Flask==3.0.0
Flask-Script==2.0.6
fonttools==4.42.1
fqdn==1.5.1
greenlet==3.0.2
h11==0.14.0
idna==3.4
inflate64==0.3.1
pip-install==1.3.5
ipykernel==6.25.2
ipynb-py-convert==0.4.6
ipython==8.15.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.1
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.2
multivolumefile==0.2.3
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.25.2
openpyxl==3.1.2
overrides==7.4.0
packaging==23.1
pandas==2.1.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
polars==0.19.2
prometheus-client==0.18.0
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2==2.9.9
pure-eval==0.2.2
py-healthcheck==1.10.1
py7zr==0.20.6
pyarrow==13.0.0
pybcj==1.0.1
pycparser==2.21
pycryptodomex==3.18.0
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.16.1
pyparsing==3.0.9
pyppmd==1.0.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
pyzstd==0.15.9
qtconsole==5.5.0
QtPy==2.4.1
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.12.0
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
stack-data==0.6.2
starlette==0.27.0
terminado==0.18.0
texttable==1.6.7
threadpoolctl==3.2.0
tinycss2==1.2.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.7
uvicorn==0.24.0.post1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.1
widgetsnbextension==4.0.9
xgboost==2.0.2
